need to fix tests

+++ route/route-APISupply.js
@@ -15,9 +15,13 @@ module.exports = function(router) {
 
     req.body.userId = req.user._id;
 
-    return new APISupply(req.body).save()
-      .then(api => res.status(201).json(api))
-      .catch(err => errorHandler(err, req, res));
+    if(req.user.isAdmin.toString() === true){
+      return new APISupply(req.body).save()
+        .then(api => res.status(201).json(api))
+        .catch(err => errorHandler(err, req, res));
+    } else {
+      errorHandler(new Error('authorization failed; user does not have access to update'), req, res);
+    }
   });
   //************GET************
   router.get('/api/newApi/getAllById/:_id', bearerAuth, (req, res) => {
@@ -36,7 +40,7 @@ module.exports = function(router) {
       .catch(err => errorHandler(err, req, res));
   });
 
-  router.get('/api/newApi/getAll', (req, res) => {
+  router.get('/api/newApi/getAll', bearerAuth, (req, res) => {
     debug('GET /api/gallery');
 
     return APISupply.find()
@@ -45,39 +49,45 @@ module.exports = function(router) {
   });
 
   //************PUT************
-  router.put('/api/newApi/:_id', bearerAuth, (req, res) => {
+  router.put('/api/newApi/:_id', bearerAuth, jsonParser, (req, res) => {
     debug('PUT /api/newApi');
 
-    return APISupply.findById(req.params._id)
-      .then(api => {
-        if(api.userId.toString() === req.user._id.toString()) {
-          api.name = req.body.name || api.name;
-          api.url = req.body.url || api.url;
-          api.desc = req.body.desc || api.desc;
-          api.examplesOfUse = req.body.examplesOfUse || api.examplesOfUse;
-          api.examplesInUse = req.body.examplesInUse || api.examplesInUse;
-          api.rating = req.body.rating || api.rating;
-          api.tokenRequired = req.body.tokenRequired || api.tokenRequired;
-          api.tokenAccessWaitTime = req.body.tokenAccessWaitTime || api.tokenAccessWaitTime;
-          api.maxReqMin = req.body.maxReqMin || api.maxReqMin;
-          api.numUsersFav = req.body.numUsersFav || api.numUsersFav;
-          api.category = req.body.category || api.category;
+    if(req.user.isAdmin.toString() === true){
+      return APISupply.findById(req.params._id)
+        .then(api => {
+          if(api.userId.toString() === req.user._id.toString()) {
+            api.name = req.body.name || api.name;
+            api.url = req.body.url || api.url;
+            api.desc = req.body.desc || api.desc;
+            api.examplesOfUse = req.body.examplesOfUse || api.examplesOfUse;
+            api.examplesInUse = req.body.examplesInUse || api.examplesInUse;
+            api.rating = req.body.rating || api.rating;
+            api.tokenRequired = req.body.tokenRequired || api.tokenRequired;
+            api.tokenAccessWaitTime = req.body.tokenAccessWaitTime || api.tokenAccessWaitTime;
+            api.maxReqMin = req.body.maxReqMin || api.maxReqMin;
+            api.numUsersFav = req.body.numUsersFav || api.numUsersFav;
+            api.category = req.body.category || api.category;
 
 
-          return api.save();
-        }
-        errorHandler(new Error('authorization failed; user does not have access to update'), req, res);
-      })
-      .then(() => res.sendStatus(204))
-      .catch(err => errorHandler(err, req, res));
+            return api.save();
+          }
+          errorHandler(new Error('authorization failed; user does not have access to update'), req, res);
+        })
+        .then(() => res.sendStatus(204))
+        .catch(err => errorHandler(err, req, res));
+    } else {
+      errorHandler(new Error('authorization failed; user does not have access to update'), req, res);
+    }
   });
+
 
   //************DELETE************
   //TODO: need to figure out how to remove apis from users favorites
   router.delete('/api/newApi/:_id', bearerAuth, (req, res) => {
     debug('DELETE /api/newApi');
 
-    return APISupply.findByIdAndRemove(req.params._id)
+    if(req.user.isAdmin.toString() === true){
+      return APISupply.findByIdAndRemove(req.params._id)
       // .then(api => {
       //   let api.favorites = x
       //   Users.findBy
@@ -85,7 +95,10 @@ module.exports = function(router) {
       //   if(api.favorites.toString() === req.user._id.toString()) return api.remove();
       //   errorHandler(new Error('authorization failed; user does not own api, and cannot delete'), req, res);
       // })
-      .then(() => res.sendStatus(204))
-      .catch(err => errorHandler(err, req, res));
+        .then(() => res.sendStatus(204))
+        .catch(err => errorHandler(err, req, res));
+    } else {
+      errorHandler(new Error('authorization failed; user does not have access to update'), req, res);
+    }
   });
 };
